move lib functions into command file. there are only 2 and they aren't large
